Debugging
Event Controller- Can't make an object that has parameters and not use parameters. I see what you were trying to do but they have to be set when you create the object. The import statement was messing an extra file as well.

EventRepo- Trying to debug. Can't use CrudRepo if it is commented out at the import